backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pytest
pytest-cov
python-multipart
python-dotenv
python-jose[cryptography]
bcrypt
arq
redis
llama-index
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-llms-deepseek
qdrant-client
llama-index-vector-stores-qdrant
EbookLib
html2text
neo4j
openai
httpx
pre-commit
fastembed

# Open Source Platform Packages
narrative-chunker
agentic-router
